Agregando funcion que elimina proyectos, snippets

+++ src/projects/projects.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, Inject, forwardRef } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Project } from "./schema/project.schema";
 import { Model } from "mongoose";
@@ -11,6 +11,7 @@ export class ProjectsService {
 	constructor(
 		@InjectModel("projects")
 		private readonly projectsModel: Model<Project>,
+		@Inject(forwardRef(()=> FoldersService))
 		private folderService: FoldersService,
 		private fileService: FilesService
 	) {}
@@ -83,16 +84,16 @@ export class ProjectsService {
 	async deleteProject(id: string): Promise<Project> {
 		const projectToBeDeleted = await this.getProjectById(id);
 
-		await this.folderService.removeProject(
-			projectToBeDeleted.parentFolder,
-			id
-		);
-
-		await Promise.all(
-			projectToBeDeleted.files.map(async (file: any) => {
-				await this.fileService.deleteFile(file);
-			})
-		);
+		if (
+			projectToBeDeleted !== null &&
+			projectToBeDeleted.files.length !== 0
+		) {
+			await Promise.all(
+				projectToBeDeleted.files.map(async (file: any) => {
+					await this.fileService.deleteFile(file);
+				})
+			);
+		}
 
 		return await this.projectsModel.findByIdAndDelete(id);
 	}
